backup_backend/requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关 - 仅支持MySQL
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
aiomysql==0.2.0

# Redis缓存
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 任务队列
celery==5.3.4
kombu==5.3.4

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 工具库
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# 生产环境不需要开发依赖

# WireGuard相关
cryptography>=41.0.0,<47.0.0
qrcode[pil]==7.4.2

# 系统监控
psutil==5.9.6

# WebSocket支持
websockets==12.0

# 邮件验证
email-validator==2.1.0
